Update README.md
The objective of this analysis is to develop a logistic regression model to predict the risk of obesity based on age and height. The dataset consists of three variables: Do_yours (an encoded representation of age), Height_cm (height in centimeters), and Danger of being fat (a binary indicator of obesity risk, where 1 denotes a risk of obesity and 0 denotes no risk). The goal is to understand how age and height influence the risk of obesity and to build a predictive model that can classify individuals into obesity risk categories.